refactor(core): San kept a second answer to a question the SDK answers
classify.go existed for one reason, and it said so at the top: to translate the
SDK's reading of a provider failure onto two interfaces of San's own, "so the
loop carries no provider vocabulary". The loop is the SDK's now and reads
ai.IsRetryable directly, so the translation had nobody left to translate for —
and the two interfaces it fed, RetryableError and ContextExceededError, were a
second vocabulary for a question already answered.

The two places outside the loop that asked it now ask the SDK. llm's one-shot
Complete keeps classifying on the way out, for the same reason it always did —
an error leaving that package is always classified, so no caller gets a
different answer by reaching for a different helper — except that what it hands
back is now an *ai.Error rather than a San wrapper around one.

retry.go goes with them, as backoff.go. A file claiming to be about retrying
was fair while it held the retry vocabulary; what is left in it is a backoff
with a cancel, for the two retry loops that are not the agent's — llm's
one-shot Complete and autopilot's steering call. Renaming it earlier would have
been renaming a file that still had retrying in it.

259 lines of vocabulary out, and 371 against 59 across the whole commit once
retry.go goes with them.
